# docs/swagger.yaml
basePath: /
definitions:
...
  models.CreatePlayerSheetRequest:
    properties:
      data:
        $ref: '#/definitions/models.PlayerSheetData'
      name:
        maxLength: 100
        minLength: 3
        type: string
      table_id:
        type: string
      template_id:
        minimum: 1
        type: integer
    required:
    - name
    - table_id
    - template_id
    type: object
  models.CreateRollRequest:
    properties:
      expression:
        maxLength: 200
        type: string
      field_name:
        maxLength: 100
        type: string
      sheet_id:
        type: string
    required:
    - sheet_id
    type: object
...
  models.PlayerSheetData:
    additionalProperties: true
    type: object
  models.PlayerSheetResponse:
    properties:
      created_at:
        type: string
      data:
        $ref: '#/definitions/models.PlayerSheetData'
      id:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/models.UserResponse'
      owner_id:
        type: integer
      table_id:
        type: string
      template:
        $ref: '#/definitions/models.TemplateInfo'
      template_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.RollDetails:
    properties:
      critical:
        description: Se foi crítico
        type: boolean
      dice:
        description: Valores individuais dos dados
        items:
          type: integer
        type: array
      fumble:
        description: Se foi fumble
        type: boolean
      modifier:
        description: Modificador aplicado
        type: integer
      total:
        description: Resultado final
        type: integer
    type: object
  models.RollResponse:
    properties:
      created_at:
        type: string
      expression:
        type: string
      field_name:
        type: string
      id:
        type: string
      result_details:
        $ref: '#/definitions/models.RollDetails'
      result_value:
        type: integer
      sheet_id:
        type: string
      success:
        type: boolean
      table_id:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
      user_id:
        type: integer
    type: object
...
  models.TemplateInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.UpdatePlayerSheetRequest:
    properties:
      data:
        $ref: '#/definitions/models.PlayerSheetData'
      name:
        maxLength: 100
        minLength: 3
        type: string
    type: object
...
  models.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: luiz@example.com
    name: Luiz de Queiroz
  description: API para gerenciamento de sessões de RPG, incluindo usuários, campanhas,
    personagens e sessões de jogo.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: RPG Session Backend API
  version: 1.0.0
paths:
...
  /api/v1/rolls:
    post:
      consumes:
      - application/json
      description: Executa rolagem de dados baseada em expressão ou campo da ficha
      parameters:
      - description: Dados da rolagem
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateRollRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RollResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Rolar dados
      tags:
      - Player Sheets
  /api/v1/rolls/sheet/{sheetID}:
    get:
      description: Lista histórico de rolagens de uma ficha específica
      parameters:
      - description: ID da ficha
        in: path
        name: sheetID
        required: true
        type: string
      - default: 1
        description: Página
        in: query
        name: page
        type: integer
      - default: 20
        description: Itens por página
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar rolagens da ficha
      tags:
      - Player Sheets
  /api/v1/rolls/table/{tableID}:
    get:
      description: Lista histórico de rolagens de uma mesa
      parameters:
      - description: ID da mesa
        in: path
        name: tableID
        required: true
        type: string
      - default: 1
        description: Página
        in: query
        name: page
        type: integer
      - default: 20
        description: Itens por página
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar rolagens da mesa
      tags:
      - Player Sheets
  /api/v1/sheets:
    get:
      description: Lista todas as fichas de personagens de uma mesa
      parameters:
      - description: ID da mesa
        in: query
        name: table_id
        required: true
        type: string
      - default: 1
        description: Página
        in: query
        name: page
        type: integer
      - default: 20
        description: Itens por página
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar fichas da mesa
      tags:
      - Player Sheets
    post:
      consumes:
      - application/json
      description: Cria nova ficha de personagem em uma mesa
      parameters:
      - description: Dados da ficha
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreatePlayerSheetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PlayerSheetResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar ficha de personagem
      tags:
      - Player Sheets
  /api/v1/sheets/{id}:
    delete:
      description: Remove uma ficha de personagem
      parameters:
      - description: ID da ficha
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remover ficha
      tags:
      - Player Sheets
    get:
      description: Busca ficha de personagem por ID
      parameters:
      - description: ID da ficha
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PlayerSheetResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buscar ficha
      tags:
      - Player Sheets
    put:
      consumes:
      - application/json
      description: Atualiza dados de uma ficha de personagem
      parameters:
      - description: ID da ficha
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualização
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePlayerSheetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PlayerSheetResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar ficha
      tags:
      - Player Sheets
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Digite ''Bearer'' seguido do token JWT (ex: Bearer eyJhbGciOiJIUzI1...)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"